feat(upgrade): update some plugins and languages

+++ public/wp-content/plugins/autoptimize/classes/autoptimizeMain.php
@@ -191,8 +191,12 @@ class autoptimizeMain
                 }
 
                 // And disable Jetpack's site accelerator if JS or CSS opt. are active.
-                if ( class_exists( 'Jetpack' ) && apply_filters( 'autoptimize_filter_main_disable_jetpack_cdn', true ) && ( $conf->get( 'autoptimize_js' ) || $conf->get( 'autoptimize_css' ) ) ) {
-                    add_filter( 'jetpack_force_disable_site_accelerator', '__return_true' );
+                if ( class_exists( 'Jetpack' ) && apply_filters( 'autoptimize_filter_main_disable_jetpack_cdn', true ) && ( $conf->get( 'autoptimize_js' ) || $conf->get( 'autoptimize_css' ) || autoptimizeImages::imgopt_active() ) ) {
+                    add_filter( 'jetpack_force_disable_site_accelerator', '__return_true' ); // this does not seemt to work any more?
+                    if ( true === autoptimizeImages::imgopt_active() ) {
+                        // only disable photon if AO is optimizing images.
+                        add_filter( 'jetpack_photon_skip_for_url', '__return_true' );
+                    }
                 }
 
                 // Add "no cache found" notice.
@@ -214,6 +218,7 @@ class autoptimizeMain
 
             // And show the imgopt notice.
             add_action( 'admin_notices', 'autoptimizeMain::notice_plug_imgopt' );
+            add_action( 'admin_notices', 'autoptimizeMain::notice_imgopt_issue' );
         }
     }
 
@@ -223,6 +228,7 @@ class autoptimizeMain
         if ( autoptimizeConfig::is_admin_and_not_ajax() ) {
             new autoptimizePartners();
             new autoptimizeExitSurvey();
+            new autoptimizeProTab();
         }
     }
 
@@ -254,10 +260,12 @@ class autoptimizeMain
 
     public function maybe_run_ao_compat()
     {
+        $conf = autoptimizeConfig::instance();
+
         // Condtionally loads the compatibility-class to ensure more out-of-the-box compatibility with big players.
         $_run_compat = true;
 
-        if ( autoptimizeOptionWrapper::get_option( 'autoptimize_installed_before_compatibility', false ) ) {
+        if ( 'on' === $conf->get( 'autoptimize_installed_before_compatibility' ) ) {
             // If AO was already running before Compatibility logic was added, don't run compat by default
             // because it can be assumed everything works and we want to avoid (perf) regressions that
             // could occur due to compatibility code.
@@ -391,9 +399,15 @@ class autoptimizeMain
             }
 
             // Misc. querystring paramaters that will stop AO from doing optimizations (pagebuilders +
-            // 2 generic parameters that could/ should become standard between optimization plugins?)
+            // 2 generic parameters that could/ should become standard between optimization plugins?).
             if ( false === $ao_noptimize ) {
-                $_qs_showstoppers = array( 'no_cache', 'no_optimize', 'tve', 'elementor-preview', 'fl_builder', 'vc_action', 'et_fb', 'bt-beaverbuildertheme', 'ct_builder', 'fb-edit', 'siteorigin_panels_live_editor', 'preview' );
+                $_qs_showstoppers = array( 'no_cache', 'no_optimize', 'tve', 'elementor-preview', 'fl_builder', 'vc_action', 'et_fb', 'bt-beaverbuildertheme', 'ct_builder', 'fb-edit', 'siteorigin_panels_live_editor', 'preview', 'td_action' );
+
+                // doing Jonathan a quick favor to allow correct unused CSS generation ;-) .
+                if ( apply_filters( 'autoptimize_filter_main_showstoppers_do_wp_rocket_a_favor', true ) ) {
+                    $_qs_showstoppers[] = 'nowprocket';
+                }
+
                 foreach ( $_qs_showstoppers as $_showstopper ) {
                     if ( array_key_exists( $_showstopper, $_GET ) ) {
                         $ao_noptimize = true;
@@ -421,6 +435,12 @@ class autoptimizeMain
             if ( function_exists( 'is_customize_preview' ) && is_customize_preview() ) {
                 $is_customize_preview = is_customize_preview();
             }
+            
+            // explicitly disable when is_login exists and is true but don't use it direclty because older versions of WordPress don't have that yet.
+            $is_login = false;
+            if ( function_exists( 'is_login' ) && true === is_login() ) {
+                $is_login = true;
+            }
 
             /**
              * We only buffer the frontend requests (and then only if not a feed
@@ -429,7 +449,7 @@ class autoptimizeMain
              * while the main query hasn't been ran yet. Thats why we use
              * AUTOPTIMIZE_INIT_EARLIER in tests.
              */
-            $do_buffering = ( ! is_admin() && ! is_feed() && ! is_embed() && ! $ao_noptimize && ! $is_customize_preview );
+            $do_buffering = ( ! is_admin() && ! is_feed() && ! is_embed() && ! $is_login && ! $is_customize_preview && ! $ao_noptimize );
         }
 
         return $do_buffering;
@@ -558,7 +578,8 @@ class autoptimizeMain
                 'minify_excluded' => $conf->get( 'autoptimize_minify_excluded' ),
             ),
             'autoptimizeHTML'    => array(
-                'keepcomments' => $conf->get( 'autoptimize_html_keepcomments' ),
+                'keepcomments'  => $conf->get( 'autoptimize_html_keepcomments' ),
+                'minify_inline' => $conf->get( 'autoptimize_html_minify_inline' ),
             ),
         );
 
@@ -618,6 +639,7 @@ class autoptimizeMain
             'autoptimize_css_exclude',
             'autoptimize_html',
             'autoptimize_html_keepcomments',
+            'autoptimize_html_minify_inline',
             'autoptimize_enable_site_config',
             'autoptimize_enable_meta_ao_settings',
             'autoptimize_js',
@@ -684,8 +706,9 @@ class autoptimizeMain
 
         // Remove AO CCSS cached files and directory.
         $ao_ccss_dir = WP_CONTENT_DIR . '/uploads/ao_ccss/';
-        if ( file_exists( $ao_ccss_dir ) && is_dir( $ao_ccss_dir ) ) {
+        if ( file_exists( $ao_ccss_dir ) && is_dir( $ao_ccss_dir ) && defined( 'GLOB_BRACE' ) ) {
             // fixme: should check for subdirs when in multisite and remove contents of those as well.
+            // fixme: if GLOB_BRACE is not avaible we need to remove AO_CCSS_DIR differently?
             array_map( 'unlink', glob( $ao_ccss_dir . '*.{css,html,json,log,zip,lock}', GLOB_BRACE ) );
             rmdir( $ao_ccss_dir );
         }
@@ -727,40 +750,64 @@ class autoptimizeMain
     {
         echo '<div class="error"><p>';
         // Translators: %s is the cache directory location.
-        printf( __( 'Autoptimize cannot write to the cache directory (%s), please fix to enable CSS/ JS optimization!', 'autoptimize' ), AUTOPTIMIZE_CACHE_DIR );
+        printf( esc_html__( 'Autoptimize cannot write to the cache directory (%s), please fix to enable CSS/ JS optimization!', 'autoptimize' ), AUTOPTIMIZE_CACHE_DIR );
         echo '</p></div>';
     }
 
     public static function notice_installed()
     {
         echo '<div class="updated"><p>';
-        printf( __( 'Thank you for installing and activating Autoptimize. Please configure it under %sSettings -> Autoptimize%s to start improving your site\'s performance.', 'autoptimize' ), '<a href="options-general.php?page=autoptimize">', '</a>' );
+        // translators: the variables contain opening and closing <a> tags to link to the settings page.
+        printf( esc_html__( 'Thank you for installing and activating Autoptimize. Your site is being optimized immediately, please test the frontend to ensure everything still works as expected. If needed you can change JavaScript or CSS optimization settings under %1$sSettings -> Autoptimize%2$s .', 'autoptimize' ), '<a href="options-general.php?page=autoptimize">', '</a>' );
         echo '</p></div>';
     }
 
     public static function notice_updated()
     {
         echo '<div class="updated"><p>';
-        _e( 'Autoptimize has just been updated. Please <strong>test your site now</strong> and adapt Autoptimize config if needed.', 'autoptimize' );
+        printf( esc_html_e( 'Autoptimize has just been updated. Please %1$stest your site now%2$s and adapt Autoptimize config if needed.', 'autoptimize' ), '<strong>', '</strong>' );
         echo '</p></div>';
     }
 
     public static function notice_plug_imgopt()
     {
         // Translators: the URL added points to the Autopmize Extra settings.
-        $_ao_imgopt_plug_notice      = sprintf( __( 'Did you know Autoptimize includes on-the-fly image optimization (with support for WebP and AVIF) and CDN via ShortPixel? Check out the %1$sAutoptimize Image settings%2$s to activate this option.', 'autoptimize' ), '<a href="options-general.php?page=autoptimize_imgopt">', '</a>' );
+        $_ao_imgopt_plug_notice      = sprintf( esc_html__( 'Did you know that Autoptimize offers on-the-fly image optimization (with support for WebP and AVIF) and CDN via ShortPixel? Check out the %1$sAutoptimize Image settings%2$s to enable this option.', 'autoptimize' ), '<a href="options-general.php?page=autoptimize_imgopt">', '</a>' );
         $_ao_imgopt_plug_notice      = apply_filters( 'autoptimize_filter_main_imgopt_plug_notice', $_ao_imgopt_plug_notice );
         $_ao_imgopt_launch_ok        = autoptimizeImages::launch_ok_wrapper();
         $_ao_imgopt_plug_dismissible = 'ao-img-opt-plug-123';
         $_ao_imgopt_active           = autoptimizeImages::imgopt_active();
         $_is_ao_settings_page        = autoptimizeUtils::is_ao_settings();
 
-        if ( current_user_can( 'manage_options' ) && $_is_ao_settings_page && '' !== $_ao_imgopt_plug_notice && ! $_ao_imgopt_active && $_ao_imgopt_launch_ok && PAnD::is_admin_notice_active( $_ao_imgopt_plug_dismissible ) ) {
+        if ( current_user_can( 'manage_options' ) && ! defined( 'AO_PRO_VERSION' ) && $_is_ao_settings_page && '' !== $_ao_imgopt_plug_notice && ! $_ao_imgopt_active && $_ao_imgopt_launch_ok && PAnD::is_admin_notice_active( $_ao_imgopt_plug_dismissible ) ) {
             echo '<div class="notice notice-info is-dismissible" data-dismissible="' . $_ao_imgopt_plug_dismissible . '"><p>';
             echo $_ao_imgopt_plug_notice;
             echo '</p></div>';
         }
     }
+
+    public static function notice_imgopt_issue()
+    {
+        // Translators: the URL added points to the Autopmize Extra settings.
+        $_ao_imgopt_issue_notice      = sprintf( esc_html__( 'Shortpixel reports it cannot always reach your site, which might mean some images are not optimized. You can %1$sread more about why this happens and how you can fix that problem here%2$s.', 'autoptimize' ), '<a href="https://shortpixel.com/knowledge-base/article/469-i-received-an-e-mail-that-says-some-of-my-images-are-not-accessible-what-should-i-do#fullarticle" target="_blank">', '</a>' );
+        $_ao_imgopt_issue_notice      = apply_filters( 'autoptimize_filter_main_imgopt_issue_notice', $_ao_imgopt_issue_notice );
+        $_ao_imgopt_issue_dismissible = 'ao-img-opt-issue-14';
+        $_ao_imgopt_active            = autoptimizeImages::imgopt_active();
+        $_ao_imgopt_status            = autoptimizeOptionWrapper::get_option( 'autoptimize_imgopt_provider_stat', '' );
+
+        if ( is_array( $_ao_imgopt_status ) && array_key_exists( 'TemporaryRedirectOrigin', $_ao_imgopt_status ) && ( $_ao_imgopt_status['TemporaryRedirectOrigin'] === "true" || $_ao_imgopt_status['TemporaryRedirectOrigin'] === true ) ) {
+            $_ao_imgopt_status_redirect_warning = true;            
+        } else {
+            $_ao_imgopt_status_redirect_warning = false;
+        }
+
+        if ( current_user_can( 'manage_options' ) && $_ao_imgopt_active && $_ao_imgopt_status_redirect_warning && '' !== $_ao_imgopt_issue_notice && PAnD::is_admin_notice_active( $_ao_imgopt_issue_dismissible ) ) {
+            echo '<div class="notice notice-info is-dismissible" data-dismissible="' . $_ao_imgopt_issue_dismissible . '"><p>';
+            echo $_ao_imgopt_issue_notice;
+            echo '</p></div>';
+        }
+    }
+
 
     public static function notice_nopagecache()
     {
@@ -769,9 +816,10 @@ class autoptimizeMain
          *
          * uses helper function in autoptimizeUtils.php
          */
-        $_ao_nopagecache_notice      = __( 'It looks like your site might not have <strong>page caching</strong> which is a <strong>must-have for performance</strong>. If you are sure you have a page cache, you can close this notice, but when in doubt check with your host if they offer this or install a page caching plugin like for example', 'autoptimize' );
-        $_ao_pagecache_install_url   = network_admin_url() . 'plugin-install.php?tab=search&type=term&s=';
-        $_ao_nopagecache_notice     .= ' <a href="' . $_ao_pagecache_install_url . 'wp+super+cache' . '">WP Super Cache</a>, <a href="' . $_ao_pagecache_install_url . 'keycdn+cache+enabler' . '">KeyCDN Cache Enabler</a>, ...';
+        // translators: strong tags and a break.
+        $_ao_nopagecache_notice      = sprintf( esc_html__( 'It looks like your site might not have %1$spage caching%2$s which is a %1$smust-have for performance%2$s. If you are sure you have a page cache, you can close this notice.%3$sWhen in doubt check with your host if they offer this or install a free page caching plugin like for example KeyCDN Cache Enabler', 'autoptimize' ), '<strong>', '</strong>', '<br />' );
+        // translators: strong tags.
+        $_ao_nopagecache_notice     .= ' ' . esc_html__('or consider ', 'autoptimize') . '<strong><a href="https://autoptimize.com/pro/">Autoptimize Pro</a></strong>' . sprintf( esc_html__( ' which not only has page caching but also image optimization, critical CSS and advanced booster options %1$sto make your site significantly faster%2$s!', 'autoptimize' ), '<strong>', '</strong>' );
         $_ao_nopagecache_dismissible = 'ao-nopagecache-forever'; // the notice is only shown once and will not re-appear when dismissed.
         $_is_ao_settings_page        = autoptimizeUtils::is_ao_settings();
 
@@ -789,8 +837,8 @@ class autoptimizeMain
         /*
          * Using other plugins to do CSS/ JS optimization can cause unexpected and hard to troubleshoot issues, warn users who seem to be in that situation.
          */
-        // Translators: the sentence will be finished with the name of the offending plugin and a final stop.
-        $_ao_potential_conflict_notice      = __( 'It looks like you have <strong>another plugin also doing CSS and/ or JS optimization</strong>, which can result in hard to troubleshoot <strong>conflicts</strong>. For this reason it is recommended to disable this functionality in', 'autoptimize' ) . ' ';
+        // Translators: some strong tags + the sentence will be finished with the name of the offending plugin and a final stop.
+        $_ao_potential_conflict_notice      = sprintf( esc_html__( 'It looks like you have %1$sanother plugin also doing CSS and/ or JS optimization%2$s, which can result in hard to troubleshoot %1$sconflicts%2$s. For this reason it is recommended to disable this functionality in', 'autoptimize' ), '<strong>', '</strong>' ) . ' ';
         $_ao_potential_conflict_dismissible = 'ao-potential-conflict-forever'; // the notice is only shown once and will not re-appear when dismissed.
         $_is_ao_settings_page               = autoptimizeUtils::is_ao_settings();
